chore: Move RequestStatusName to codecs
