+++ packages/anchor-utils/src/index.ts
@@ -1,5 +1,4 @@
 export * from './cid.js'
 export * from './constants.js'
 export * from './interface.js'
-export * from './request-status-name.js'
 export * from './merkle.js'
